Bid Value updated to Float

+++ app/src/main/java/com/CioffiDeVivo/dietideals/DataModels/Bid.kt
@@ -5,12 +5,12 @@ import java.util.UUID
 
 class Bid(
     id: UUID,
-    value: Int,
+    value: Float,
     userId: UUID,
     date: ZonedDateTime
 ) {
     val id: UUID = id
-    val value: Int = value
+    val value: Float = value
     val userId: UUID = userId
     val date: ZonedDateTime = date
 }
